Implemented High Score Record, Lives, Game Over Screen and Game Restart
Created the Game Over Screen. This contains the High Score and a button to restart the game.

Created C# script LivesController which is responsible for decreasing player lives by one every time the snake dies and displaying it.

Created C# script and game object for Game Controller which is responsible for displaying game over screen, calling the function to set high score, disabling consumable script on when there are no more lives left and restart the game on clicking the restart button on the Game Over Screen.

Used Singleton principle and created C# script and Game object for HighScoreController and defined it as an object which persists across multiple game sessions. This is responsible for storing the player's high score.

Made changes to the Snake C# script where the script is disabled when there are no player lives left.

Created C# script HighScoreDisplay to retrieve and display Player's high score on the Game Over screen.