Implemented Powerups
Created C# script Powerup Controller which is responsible for spawning and despawning the different Powerups.

Implemented two different powerups: Shield and Score Boost.

Made changes to Snake C# script by adding the amount of time the powerup lasts and specifying the effects each powerup has on the snake. Added function for Disabling powerup when time of powerup effect is up.