Prevented snake from crossing walls when using shield powerup
Made changes to the Snake C# script by making the snake go in the direction opposite to what it was going when it collides with the wall while using shield powerup hence preventing it from crossing the wall.

Increased the Powerup effect time from 3sec to 5sec.

Added new tags Wall and SnakeBody and assigned it to walls and snake segments respectively.