Added a new Game Mode (Player 1 vs Player 2) and made corrections
Created a new Scene for Player 1 vs Player 2 game mode.

Changed the Game UI by removing score and adding Player 1 Lives and Player 2 Lives.

Added a new Win Result text mesh game object to Game over Screen. Created C# script WinResult which compares the lives of each player and determines which player has won or if it is a draw.

Added a button to the Lobby Screen and Added function to LobbyController to load the Vs game mode.

Made changes to Game Controller so that Win result only displays in the Vs game mode and not others.