Implemented Speed Boost Powerup
Added image asset for Speed Boost powerup.

Created new Game Object for Speed Boost powerup.

Made changes to the Powerup C# script by adding the by adding variable to store Speed boost power up game object and add it to the list of powerup game objects to be spawned.

Made changes to the SnakeController C# script by specifying the effect of Speed Boost power up when snake collides with it.

Made changes to the PowerupStatus C# script by specifying the color of the notification to be displayed when the speed boost power up is consumed.